pyproject.toml
[project]
name = "torch-learning"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "matplotlib>=3.10.8",
    "mlxtend>=0.24.0",
    "pandas>=3.0.1",
    "pillow>=12.1.1",
    "requests>=2.32.5",
    "scikit-learn>=1.8.0",
    "torch>=2.10.0",
    "torchmetrics>=1.9.0",
    "torchsummary>=1.5.1",
    "torchvision>=0.25.0",
    "tqdm>=4.67.3",
]
